refactor(meshtastic): improve type hinting and update meshtastic imports
Update `create_ble_interface` to use `Callable` and `object` for better type precision. Update multiple test files to import `BROADCAST_NUM` directly from `meshtastic` instead of `meshtastic.mesh_interface`.

Additionally, clean up `TestWeatherPlugin` by removing redundant mocks for `matches` and `get_matching_matrix_command_with_args`, and update test cases to use `mock_event.body` for more realistic message handling.
