+++ tests/test_meshtastic_utils_edge_cases.py
@@ -21,7 +21,7 @@ from concurrent.futures import TimeoutError as ConcurrentTimeoutError
 from typing import NoReturn
 from unittest.mock import ANY, AsyncMock, MagicMock, Mock, patch
 
-from meshtastic.mesh_interface import BROADCAST_NUM
+from meshtastic import BROADCAST_NUM
 
 # Add src to path for imports
 sys.path.insert(0, os.path.join(os.path.dirname(__file__), "..", "src"))
